adds restructuring, layout, fixes marquee scroll
