+++ src/components/builder/characterStats/index.js
@@ -1,8 +1,8 @@
 import { useState } from 'react';
-import StatsEdit from 'components/statEdit';
-import QuestModifiers from 'components/questModifiers';
-import ElectiveModifiers from 'components/electiveModifiers';
-import CharacterFinal from 'components/characterFinal';
+import StatsEdit from 'components/builder/statEdit';
+import QuestModifiers from 'components/builder/questModifiers';
+import ElectiveModifiers from 'components/builder/electiveModifiers';
+import CharacterFinal from 'components/builder/characterFinal';
 import { makeStyles } from '@material-ui/styles';
 
 const useStyles = makeStyles({
