+++ src/components/builder/equipmentList/index.js
