+++ src/components/common/detail/detailFieldLabel/index.js
