+++ src/components/common/detail/detailHeader/index.js
