+++ src/components/common/detail/index.js
