+++ src/components/common/hubTabs/tabDetail/index.js
