+++ src/components/main/background/index.js
