+++ src/layouts/hub.js
@@ -1,19 +1,27 @@
-import Background from 'components/background';
-import NavBar from 'components/hubNavbar/hubNavBar';
-import Toolbar from 'components/toolbar';
+import { PropTypes } from 'prop-types';
+import Background from 'components/main/background';
+import NavBar from 'components/main/hubNavbar/hubNavBar';
+import Toolbar from 'components/common/toolbar';
 
 const Hub = props => {
   const { children } = props;
 
   return (
-    <div style={{ display: 'flex', flexDirection: 'column', flex: 1, overflow: 'hidden' }}>
+    <div style={{
+      display: 'flex', flexDirection: 'column', flex: 1, overflow: 'hidden',
+    }}
+    >
       <Background>
         <NavBar />
         <Toolbar />
         {children}
       </Background>
     </div>
-  )
+  );
+};
+
+Hub.propTypes = {
+  children: PropTypes.node.isRequired,
 };
 
 export default Hub;
